Finish refactor
Removed all duplicate code, moved to a single Screen file (bumped version to reflect this) and changed translations to better support the new system.
