+++ gradle.properties
@@ -6,7 +6,7 @@ minecraft_version=1.18.2
 yarn_mappings=1.18.2+build.3
 loader_version=0.13.3
 # Mod Properties
-mod_version=1.0.1
+mod_version=2.0.0
 maven_group=com.p4
 archives_base_name=cycleswitcher
 # Dependencies
