+++ src/main/resources/assets/cycleswitcher/lang/en_us.json
@@ -2,27 +2,27 @@
   "weather.clear": "Clear",
   "weather.rain": "Rain",
   "weather.thunder": "Thunder",
-  "debug.weathers.select_next": "%s Next",
-  "debug.weathers.press_f5": "[ F5 ]",
-  "debug.weathers.help": "F3 + F5 = Open weather switcher",
-  "debug.weathers.error": "Unable to open weather switcher; no permission",
+  "debug.weather.select_next": "%s Next",
+  "debug.weather.press_key": "[ F5 ]",
+  "debug.weather.help": "F3 + F5 = Open weather switcher",
+  "debug.weather.error": "Unable to open weather switcher; no permission",
 
   "time.day": "Day",
   "time.noon": "Noon",
   "time.night": "Night",
   "time.midnight": "Midnight",
-  "debug.times.select_next": "%s Next",
-  "debug.times.press_f6": "[ F6 ]",
-  "debug.times.help": "F3 + F6 = Open time switcher",
-  "debug.times.error": "Unable to open time switcher; no permission",
+  "debug.time.select_next": "%s Next",
+  "debug.time.press_key": "[ F6 ]",
+  "debug.time.help": "F3 + F6 = Open time switcher",
+  "debug.time.error": "Unable to open time switcher; no permission",
 
   "difficulty.peaceful": "Peaceful",
   "difficulty.easy": "Easy",
   "difficulty.normal": "Normal",
   "difficulty.hard": "Hard",
-  "debug.difficulties.select_next": "%s Next",
-  "debug.difficulties.press_f7": "[ F7 ]",
-  "debug.difficulties.help": "F3 + F7 = Open difficulty switcher",
-  "debug.difficulties.error": "Unable to open difficulty switcher; no permission",
-  "debug.difficulties.error2": "Unable to open difficulty switcher; difficulty locked"
+  "debug.difficulty.select_next": "%s Next",
+  "debug.difficulty.press_key": "[ F7 ]",
+  "debug.difficulty.help": "F3 + F7 = Open difficulty switcher",
+  "debug.difficulty.error": "Unable to open difficulty switcher; no permission",
+  "debug.difficulty.error2": "Unable to open difficulty switcher; difficulty locked"
 }
